refactor(ui): Notification no longer gets added if already on the activePane
Just a small thing I noticed. I see this mainly happen on the Technology pane, but any time an unlock occurs for the pane that is already active, the notification badge still appears. This change prevents the notification from being added in the first place if you're already on that pane.